requirements.txt
google-generativeai
feedparser